Replace card button with header button
The add to library card button was replaced with a statically positioned
button in the header to improve the UX. Having the button on a card
and dynamically move as books are added to the library would frustrate
the user if they constantly need to locate where the button is. The new
button is also more intuitive to use.