Add aggressive validation to form
Originally, it wasn't added due to low criteria requirements for the
form fields. It is being added in this commit to improve the UX of the
form and solve a bug where submitting the form with an invalid input
requires the user to submit the form twice: once to update the
validation status of the form and second to actually submit the form.