Fix bug where submitting edit causes page to crash
The crash was due to picking the wrong element children for a card due
to the addition of containers for the card buttons and book information
containing elements.

The flavor text for the pages element on a card was also updated to
match what the form requests the user to input.